- finished reorganization, ready for testing - removed all unnecessary meteor packages
