+++ .meteor/versions
@@ -1,36 +1,23 @@
-accounts-base@1.2.11
-accounts-password@1.3.0
-alanning:roles@1.2.15
-aldeed:autoform@5.5.1
 aldeed:collection2@2.10.0
 aldeed:collection2-core@1.2.0
 aldeed:schema-deny@1.1.0
 aldeed:schema-index@1.1.0
 aldeed:simple-schema@1.5.3
-aldeed:tabular@1.4.2
-aldeed:template-extension@4.0.0
 allow-deny@1.0.5
-anti:i18n@0.4.3
 autoupdate@1.2.11
 babel-compiler@6.9.1
 babel-runtime@0.1.11
 base64@1.0.9
-bevanhunt:leaflet@2.0.1
 binary-heap@1.0.9
 blaze@2.1.8
-blaze-html-templates@1.0.4
 blaze-tools@1.0.9
 boilerplate-generator@1.0.9
-caching-compiler@1.1.6
-caching-html-compiler@1.0.6
 callback-hook@1.0.9
 check@1.2.3
-coffeescript@1.0.17
 dburles:collection-helpers@1.0.4
 ddp@1.2.5
 ddp-client@1.2.9
 ddp-common@1.2.6
-ddp-rate-limiter@1.0.5
 ddp-server@1.2.10
 deps@1.0.12
 diff-sequence@1.0.6
@@ -40,95 +27,51 @@ ejson@1.0.12
 email@1.1.17
 epaminond:nodemailer@1.4.0
 es5-shim@4.6.14
-fastclick@1.0.12
 fongandrew:find-and-modify@1.0.0
-fortawesome:fontawesome@4.5.0
 geojson-utils@1.0.9
 harrison:papa-parse@1.1.7
 hot-code-push@1.0.4
 html-tools@1.0.10
 htmljs@1.0.10
 http@1.1.8
-ian:accounts-ui-bootstrap-3@1.2.89
 id-map@1.0.8
-iron:controller@1.0.12
-iron:core@1.0.11
-iron:dynamic-template@1.0.12
-iron:layout@1.0.12
-iron:location@1.0.11
-iron:middleware-stack@1.1.0
-iron:router@1.0.13
-iron:url@1.0.11
-johdirr:meteor-git-rev@0.0.4
 jquery@1.11.9
-juliancwirko:s-alert@3.2.0
-launch-screen@1.0.12
-less@2.5.7
 livedata@1.0.18
-localstorage@1.0.11
 logging@1.1.15
-makalu:better-dateinput-polyfill@0.0.2
-mdg:geolocation@1.3.0
 mdg:validation-error@0.2.0
 meteor@1.2.17
 meteor-base@1.0.4
 meteorhacks:aggregate@1.3.0
 meteorhacks:collection-utils@1.2.0
-meteorhacks:meteorx@1.4.1
-meteorhacks:unblock@1.1.0
-mfactory:admin-lte@0.0.2
 minifier-css@1.2.14
 minifier-js@1.2.14
 minimongo@1.0.17
-mobile-experience@1.0.4
-mobile-status-bar@1.0.12
 modules@0.7.6
 modules-runtime@0.7.6
 momentjs:moment@2.14.4
 mongo@1.1.12
 mongo-id@1.0.5
 mongo-livedata@1.0.12
-msavin:sjobs@2.0.1
-npm-bcrypt@0.9.1
 npm-mongo@1.5.48
 observe-sequence@1.0.12
 ordered-dict@1.0.8
-peppelg:bootstrap-3-modal@1.0.4
 promise@0.8.4
 raix:eventemitter@0.1.3
-raix:handlebar-helpers@0.2.5
 random@1.0.10
-rate-limit@1.0.5
-reactive-dict@1.1.8
 reactive-var@1.0.10
 reload@1.1.10
 retry@1.0.8
-reywood:publish-composite@1.4.2
 routepolicy@1.0.11
-service-configuration@1.0.10
-session@1.1.6
-sha@1.0.8
 shell-server@0.2.1
-softwarerero:accounts-t9n@1.3.4
 spacebars@1.0.12
 spacebars-compiler@1.0.12
-srp@1.0.9
 standard-minifier-css@1.2.0
 standard-minifier-js@1.2.0
-stylus@2.513.4
-templating@1.1.14
-templating-tools@1.0.4
 tracker@1.1.0
-twbs:bootstrap@3.3.6
 udondan:bulk-collection-update@0.2.0
 ui@1.0.11
 underscore@1.0.9
 url@1.0.10
-useraccounts:core@1.14.2
-useraccounts:iron-routing@1.14.2
 votercircle:winston@0.0.3
 webapp@1.3.11
 webapp-hashing@1.0.9
-williamledoux:bootstrap-toggle@2.2.1
-yogiben:admin@1.2.8
-zimme:active-route@2.3.2
