+++ imports/cronjobs/cronJob_push.js
@@ -1,8 +1,13 @@
-import '../methods/pushsitedata';
+import { Meteor } from 'meteor/meteor';
+import { pushSiteData } from '../methods/pushSiteDataFunctions';
+import { LiveSites } from '../api/collections_server';
 
-// every 10 mins push data
+// push data if site is active
 Meteor.setInterval(() => {
-  Meteor.call('pushSiteData',
-  // reading aqsid from environment (*.json pm2 config file)
-  process.env.aqsid, false);
-}, 10 * 60 * 1000); // run every 10 min, to push new data
+  // get site reading aqsid from environment (*.json pm2 config file)
+  const site = LiveSites.findOne({ AQSID: process.env.aqsid });
+
+  if (site.status === 'Active') {
+    pushSiteData(process.env.aqsid, false);
+  }
+}, 10 * 60 * 1000); // run every 10 min
