+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "dash_data_maps",
+  "name": "hnet_data_maps_server",
   "version": "0.9.0",
-  "description": "HNET DataMaps - a real time sensor analysis & visualization frame work",
+  "description": "DASH HNET DataMaps - a real time analysis & visualization frame work for sensor data",
   "main": "main.html",
   "dependencies": {
     "bcrypt": "^0.8.7",
@@ -17,33 +17,41 @@
     "lint": "eslint .",
     "pretest": "npm run lint --silent"
   },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/DataAnalyticsinStudentHands/HNETDataMaps.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/DataAnalyticsinStudentHands/DataMaps/issues"
+  },
+  "homepage": "https://github.com/DataAnalyticsinStudentHands/DataMaps",
+  "devDependencies": {
+    "babel-eslint": "^6.1.2",
+    "eslint": "^3.5.0",
+    "eslint-config-airbnb": "^11.1.0",
+    "eslint-import-resolver-meteor": "^0.3.3",
+    "eslint-plugin-import": "^1.15.0",
+    "eslint-plugin-jsx-a11y": "^2.2.2",
+    "eslint-plugin-meteor": "^4.0.0",
+    "eslint-plugin-react": "^6.2.2"
+  },
   "eslintConfig": {
-    "env": {
-      "browser": true,
-      "meteor": true
+    "parser": "babel-eslint",
+    "parserOptions": {
+      "allowImportExportEverywhere": true
     },
     "plugins": [
       "meteor"
     ],
     "extends": [
-      "airbnb/base",
-      "plugin:meteor/guide"
+      "airbnb",
+      "plugin:meteor/recommended"
     ],
-    "globals": {
-      "chokidar": true,
-      "Nodemailer": true,
-      "moment": true,
-      "logger": true,
-      "Papa": true,
-      "sAlert": true,
-      "DataSeries": true,
-      "DataExporter": true,
-      "SimpleSchema": true,
-      "channelHash": true,
-      "flagsHash": true,
-      "unitsHash": true
-    },
     "rules": {
+      "import/no-extraneous-dependencies": "off",
+      "no-underscore-dangle": "off",
       "max-len": [
         2,
         160,
@@ -57,15 +65,6 @@
       "arrow-body-style": [
         "error",
         "always"
-      ],
-      "no-underscore-dangle": [
-        "error",
-        {
-          "allow": [
-            "foo_",
-            "_id"
-          ]
-        }
       ],
       "comma-dangle": [
         "error",
@@ -89,27 +88,9 @@
           "allowForLoopAfterthoughts": true
         }
       ]
+    },
+    "settings": {
+      "import/resolver": "meteor"
     }
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/DataAnalyticsinStudentHands/HNETDataMaps.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/DataAnalyticsinStudentHands/HNETDataMaps/issues"
-  },
-  "homepage": "https://github.com/DataAnalyticsinStudentHands/HNETDataMaps",
-  "devDependencies": {
-    "@meteorjs/eslint-config-meteor": "^1.0.5",
-    "babel-eslint": "^7.2.3",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-import-resolver-meteor": "^0.4.0",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-meteor": "^4.0.1",
-    "eslint-plugin-react": "^7.0.1"
   }
 }
